+++ computer_mode.h
@@ -3,6 +3,6 @@
 
 #include "computer_mode.c"
 
-void computersMove(int *computer_move, int height, int width, char board[][width], char symbol);
+void computersMove(int *computer_move, unsigned long long height, unsigned long long width, char board[][width], char symbol);
 
 #endif
