+++ game.h
@@ -7,7 +7,7 @@
 /// @param width  The width of the board
 /// @param board  The board of the game
 /// @return Returns one if the game board is full, returns zero if the game board is not full.
-int isFull(int height, int width, char board[][width]);
+int isFull(unsigned long long height, unsigned long long width, char board[][width]);
 
 
 /// @brief  Function to check if the column chose by the user is avaliable, if not avaliable,
@@ -17,7 +17,7 @@ int isFull(int height, int width, char board[][width]);
 /// @param width  The width of the board
 /// @param board  The board of the game 
 /// @return Returns one if the column is avaliable, returns zero if the column is not avaliable
-int isColumnAvaliable(int move, int height, int width, char board[][width]);
+int isColumnAvaliable(int move, unsigned long long height, unsigned long long width, char board[][width]);
 
 /// @brief        Function edits the board to the state of the board after the user puts the piece in the avaliable place
 /// @param move   The index of the column chosen by the user
@@ -30,7 +30,7 @@ int isColumnAvaliable(int move, int height, int width, char board[][width]);
 /// @param undo    checks if the player made undo or not at that move, 1 if undo, 2 if redo, 3 at redo error
 /// @param redos_stack  sends the stack of redos
 /// @param count_redos  sends the count of undos in redos 
-void dotheMove(int move, int height, int width, char board[][width], char symbol, int moves_stack[width * height], int *counter, int *undo, int redos_stack[], int *count_redos, int mode);
+void dotheMove(int move, unsigned long long height, unsigned long long width, char board[][width], char symbol, int moves_stack[width * height], int *counter, int *undo, int redos_stack[], int *count_redos, int mode);
 
 /// @brief        Function that deals with player's moves each turn 
 ///               and it will be branched to multiple main functions until the game ends
@@ -40,7 +40,7 @@ void dotheMove(int move, int height, int width, char board[][width], char symbol
 /// @param p1     Player One
 /// @param p2     Player Two (if the user chose vs Player mode)
 /// @param computer Computer (if the user chose vs Computer mode)
-void play(int height, int width, char board[][width], player p1, player p2, player computer, int mode);
+void play(unsigned long long height, unsigned long long width, char board[][width], player p1, player p2, player computer, int mode);
 
 
 /// @brief Function that branches the player to two functions -> playVSHuman or -> playVSComputer
@@ -52,6 +52,6 @@ void play(int height, int width, char board[][width], player p1, player p2, play
 /// @param p1     Player One
 /// @param p2     Player Two (if the user chose vs Player mode)
 /// @param computer Computer (if the user chose vs Computer mode)
-void chooseMode(int game_mode, int height,int width, char board[][width], player p1, player p2, player computer);
+void chooseMode(int game_mode, unsigned long long height, unsigned long long width, char board[][width], player p1, player p2, player computer);
 
 #endif
