+++ ingamemenu.h
@@ -13,7 +13,7 @@
 /// @param undo_redo variable to indicate if you the user choose to undo or redo -> it is set 1 if he undos, 0 if he redos
 /// @param redos_stack  sends the stack of redos
 /// @param count_redos  sends the count of undos in redos 
-void undoOrRedo(int move, int height, int width, char board[][width], int moves_stack[width * height], int *counter, int *undo, int undo_redo, int redos_stack[], int *count_redos, int mode);
+void undoOrRedo(int move, unsigned long long height, unsigned long long width, char board[][width], int moves_stack[width * height], int *counter, int *undo, int undo_redo, int redos_stack[], int *count_redos, int mode);
 
 /// @brief  Function prints in-game menu and deals with option of the user
 /// @param move   The index of the column chosen by the user
@@ -25,6 +25,6 @@ void undoOrRedo(int move, int height, int width, char board[][width], int moves_
 /// @param undo    checks if the player made undo or not at that move, 1 if undo, 2 if redo, 3 at redo error
 /// @param redos_stack  sends the stack of redos
 /// @param count_redos  sends the count of undos in redos 
-void inGameMenu(int move, int height, int width, char board[][width], int moves_stack[width * height], int *counter, int *undo, int redos_stack[], int *count_redos, int mode);
+void inGameMenu(int move, unsigned long long height, unsigned long long width, char board[][width], int moves_stack[width * height], int *counter, int *undo, int redos_stack[], int *count_redos, int mode);
 
 #endif
